requirements.txt
# -f https://download.pytorch.org/whl/torch_stable.html
pytorch
jupyter
matplotlib
pandas
flask